Address a couple TODOs and fix coloration of "durability" bars

+++ src/api/java/moze_intel/projecte/api/capabilities/item/IItemCharge.java
@@ -29,8 +29,8 @@ public interface IItemCharge {
 	 *
 	 * @return The percent charge on the stack
 	 */
-	default double getChargePercent(@Nonnull ItemStack stack) {
-		return (double) getCharge(stack) / getNumCharges(stack);
+	default float getChargePercent(@Nonnull ItemStack stack) {
+		return (float) getCharge(stack) / getNumCharges(stack);
 	}
 
 	/**
